Move users to separate file
