+++ website/siteConfig.js
@@ -8,17 +8,7 @@
 // See https://docusaurus.io/docs/site-config for all the possible
 // site configuration options.
 
-// List of projects/orgs using your project for the users page.
-const users = [
-  {
-    caption: "Domapic",
-    // You will need to prepend the image path with your baseUrl
-    // if it is not '/', like: '/test-site/img/image.jpg'.
-    image: "https://domapic.com/assets/domapic-logo.png",
-    infoLink: "https://www.domapic.com",
-    pinned: false
-  }
-];
+const users = require("./users");
 
 const organizationUrl = "https://github.com/mocks-server";
 const baseUrl = "/";
