+++ website/users.js
@@ -1,0 +1,14 @@
+// List of projects/orgs using your project for the users page.
+
+const users = [
+  {
+    caption: "Domapic",
+    // You will need to prepend the image path with your baseUrl
+    // if it is not '/', like: '/test-site/img/image.jpg'.
+    image: "https://domapic.com/assets/domapic-logo.png",
+    infoLink: "https://www.domapic.com",
+    pinned: false
+  }
+];
+
+module.exports = users;
